Add support for region restrictions.
Here we add support for region restrictions. Unfortunately, the algorithm is
going to need to be a bit more complicated to account for cases where the
crop-region is a corner, but the user selects a side. It might mean this
is the wrong approach.

In any case, the flags also support arrays and booleans now.